Casestudy file update
